pyproject.toml
[tool.poetry]
authors = ["coolworld2049 <coolworld2049@gmail.com>"]
description = ""
name = "schedule_service"
readme = "README.md"
version = "1.4.8"

[tool.poetry.dependencies]
apscheduler = "^3.10.4"
beautifulsoup4 = "^4.12.2"
cashews = { extras = ["redis"], version = "^6.3.0" }
fastapi = "^0.104.0"
gunicorn = "^21.2.0"
httptools = "^0.6.0"
loguru = "^0.7.0"
openpyxl = "^3.1.2"
prometheus-client = "^0.17.0"
prometheus-fastapi-instrumentator = "^6.1.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python = "^3.11"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
redis = "^5"
requests = "^2.31.0"
tenacity = "^8.2.3"
toml = "^0.10.2"
twine = "^4.0.2"
typing-extensions = "^4.7.1"
ujson = "^5.8.0"
uvicorn = "^0.23.2"
yarl = "^1.9.2"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-env = "^1.0.1"
httpx = "^0.25.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
testpaths = [
    "schedule_service/services/vuc_schedule_parser/tests",
    "schedule_service/services/workbook_updater/tests",
]
addopts = ["--cov=./schedule_service", "--ignore=schedule_service/pypi"]

[tool.coverage.run]
source = ["./schedule_service"]
omit = ["dev/*", "tests/*", "pypi/*"]
